+++ common/src/main/java/com/faboslav/friendsandfoes/client/render/entity/renderer/MaulerEntityRenderer.java
@@ -6,6 +6,7 @@ import com.faboslav.friendsandfoes.entity.MaulerEntity;
 import com.faboslav.friendsandfoes.init.FriendsAndFoesEntityModelLayer;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
+import net.minecraft.client.render.VertexConsumerProvider;
 import net.minecraft.client.render.entity.EntityRendererFactory;
 import net.minecraft.client.render.entity.MobEntityRenderer;
 import net.minecraft.client.util.math.MatrixStack;
@@ -15,13 +16,28 @@ import net.minecraft.util.Identifier;
 @SuppressWarnings({"rawtypes", "unchecked"})
 public final class MaulerEntityRenderer extends MobEntityRenderer<MaulerEntity, MaulerEntityModel<MaulerEntity>>
 {
+	private static final float SHADOW_RADIUS = 0.35F;
+
 	public MaulerEntityRenderer(EntityRendererFactory.Context context) {
-		super(context, new MaulerEntityModel(context.getPart(FriendsAndFoesEntityModelLayer.MAULER_LAYER)), 0.35F);
+		super(context, new MaulerEntityModel(context.getPart(FriendsAndFoesEntityModelLayer.MAULER_LAYER)), SHADOW_RADIUS);
 	}
 
 	@Override
 	public Identifier getTexture(MaulerEntity mauler) {
 		return FriendsAndFoes.makeID("textures/entity/mauler/" + mauler.getMaulerType().getName() + ".png");
+	}
+
+	@Override
+	public void render(
+		MaulerEntity mauler,
+		float f,
+		float tickDelta,
+		MatrixStack matrixStack,
+		VertexConsumerProvider vertexConsumerProvider,
+		int i
+	) {
+		this.shadowRadius = mauler.isBurrowedDown() ? 0.0F:SHADOW_RADIUS;
+		super.render(mauler, f, tickDelta, matrixStack, vertexConsumerProvider, i);
 	}
 
 	@Override