+++ common/src/main/java/com/faboslav/friendsandfoes/platform/TotemHelper.java
@@ -3,18 +3,30 @@ package com.faboslav.friendsandfoes.platform;
 import com.faboslav.friendsandfoes.FriendsAndFoes;
 import dev.architectury.injectables.annotations.ExpectPlatform;
 import net.minecraft.entity.LivingEntity;
+import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.Identifier;
+import org.jetbrains.annotations.Nullable;
 
-public final class TotemPacketHelper
+import java.util.function.Predicate;
+
+public final class TotemHelper
 {
 	public static final Identifier TOTEM_EFFECT_PACKET = FriendsAndFoes.makeID("totem_effect_packet");
+	public static final String TRINKETS_MOD_ID = "trinkets";
+	public static final String CURIOS_MOD_ID = "curios";
 
 	@ExpectPlatform
 	public static void sendTotemEffectPacket(ItemStack itemStack, LivingEntity livingEntity) {
 		throw new AssertionError();
 	}
 
-	private TotemPacketHelper() {
+	@ExpectPlatform
+	@Nullable
+	public static ItemStack getTotemFromModdedSlots(PlayerEntity player, Predicate<ItemStack> totemFilter) {
+		throw new AssertionError();
+	}
+
+	private TotemHelper() {
 	}
 }