+++ forge/src/main/java/com/faboslav/friendsandfoes/network/PacketHandler.java
@@ -1,6 +1,6 @@
 package com.faboslav.friendsandfoes.network;
 
-import com.faboslav.friendsandfoes.platform.TotemPacketHelper;
+import com.faboslav.friendsandfoes.platform.TotemHelper;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraftforge.network.NetworkDirection;
@@ -12,7 +12,7 @@ public final class PacketHandler
 {
 	private static final String PROTOCOL_VERSION = "1";
 
-	public static final SimpleChannel TOTEM_CHANNEL = NetworkRegistry.newSimpleChannel(TotemPacketHelper.TOTEM_EFFECT_PACKET, () -> PROTOCOL_VERSION, PROTOCOL_VERSION::equals, PROTOCOL_VERSION::equals);
+	public static final SimpleChannel TOTEM_CHANNEL = NetworkRegistry.newSimpleChannel(TotemHelper.TOTEM_EFFECT_PACKET, () -> PROTOCOL_VERSION, PROTOCOL_VERSION::equals, PROTOCOL_VERSION::equals);
 
 	public static void registerMessages() {
 		TOTEM_CHANNEL.registerMessage(0, TotemEffectPacket.class, TotemEffectPacket::encode, TotemEffectPacket::new, TotemEffectPacket::handle);