+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## mc1.19.3-1.8.3, June 21, 2023
+
+- Added curios/trinkets totems compatibility
+- Fixed mauler infinite experience glitch
+- Fixed maulers shadow rendering when burrowed down
+- 
 ## mc1.19.3-1.8.2, May 17, 2023
 
 - Added Totem of Illusion
@@ -50,6 +56,12 @@
 ## mc1.19.3-1.7.0, December 14, 2022
 
 - Ported to 1.19.3
+
+## mc1.19.2-1.8.3, June 21, 2023
+
+- Added curios/trinkets totems compatibility
+- Fixed mauler infinite experience glitch
+- Fixed maulers shadow rendering when burrowed down
 
 ## mc1.19.2-1.8.2, May 17, 2023